refactor auth out of services, it is a bit of different
